Refactored everything
